Fix async components diffing
We have to define the class `constructor`, else async components are
recreated everytime the parent rerenders.

Technical detail: if not defined on the `Async` class, the property
`constructor` is inherited from `Component`, and preact does not find
the actual `Async` constructor in `buildComponentFromVNode`, so it
doesn't find the component owner, so the component is recreated.