Compare prerender URL with normalized pathname
Currently `entry.js` includes logic to check whether the current URL has been prerendered and can be hydrated. The comparison is between a normalized url (appended with a `\`) and the current `location.pathname`.

That means that in a situation where `/user/` has been prerendered and the client requests `/user` it will not hydrate but instead render the entire page.

Normalizing the `location.pathname` just like the prerender url will solve this.