Introduce custom hooks
They are useful for other plugins to tap into certain points of the
build process. In this case a `devServerRunning` hook is introduced,
which would allow a 3rd party plugin to, for example, display a QR code
with the address the app is running on once the dev server started.