docs: Add mention of widget-typescript template
I noticed that there is a `widget-typescript` template among the official templates.

I don't know if it's ready for general use yet, but here's a PR to update the docs for whenever it is the right time.

I also added periods to the other items in the list for consistency.